# front_start/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API для управления квантовыми задачами и гейтами",
        "title": "Quantum Tasks API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/quantum_tasks": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Возвращает список квантовых задач. Пользователь видит только свои задачи, модератор - все задачи, неавторизованный пользователь - ошибку доступа.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "QuantumTasks"
                ],
                "summary": "Получить список задач",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Фильтр по статусу",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Начальная дата (фильтр от)",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Конечная дата (фильтр до)",
                        "name": "to",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список задач",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/internal_app_handler.DTO_Resp_Tasks"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/quantum_tasks/{id}": {
            "get": {
                "description": "Возвращает детальную информацию о квантовой задаче по её идентификатору",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "QuantumTasks"
                ],
                "summary": "Получить задачу по ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID задачи",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Детали задачи",
                        "schema": {
                            "$ref": "#/definitions/internal_app_handler.DTO_Resp_Tasks"
                        }
                    },
                    "400": {
                        "description": "Invalid task ID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Task not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "description": "Обновляет описание квантовой задачи",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "QuantumTasks"
                ],
                "summary": "Обновить задачу",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID задачи",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Данные для обновления",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_app_handler.DTO_Req_TaskUpd"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Обновленная задача",
                        "schema": {
                            "$ref": "#/definitions/internal_app_handler.DTO_Resp_Tasks"
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "Полностью удаляет квантовую задачу из системы",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "QuantumTasks"
                ],
                "summary": "Удалить задачу",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID задачи",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ID удаленной задачи",
                        "schema": {
                            "$ref": "#/definitions/internal_app_handler.DTO_Resp_SimpleID"
                        }
                    },
                    "400": {
                        "description": "Invalid task ID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/quantum_tasks/{id}/form": {
            "put": {
                "description": "Переводит задачу из статуса черновика в статус сформированной",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "QuantumTasks"
                ],
                "summary": "Сформировать задачу",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID задачи",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Сформированная задача",
                        "schema": {
                            "$ref": "#/definitions/internal_app_handler.DTO_Resp_Tasks"
                        }
                    },
                    "400": {
                        "description": "Invalid task ID or cannot form task",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/quantum_tasks/{id}/resolve": {
            "put": {
                "description": "Выполняет завершение или отклонение квантовой задачи с вычислением результата",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "QuantumTasks"
                ],
                "summary": "Завершить/отклонить задачу",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID задачи",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Действие с задачей",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_app_handler.DTO_Req_TaskResolve"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Решенная задача",
                        "schema": {
                            "$ref": "#/definitions/internal_app_handler.DTO_Resp_Tasks"
                        }
                    },
                    "400": {
                        "description": "Invalid input or missing required fields",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Task not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/tasks/{task_id}/services/{service_id}": {
            "put": {
                "description": "Обновляет значение градусов для конкретного гейта в задаче",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "M-M"
                ],
                "summary": "Обновить градусы гейта",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID задачи",
                        "name": "task_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID гейта",
                        "name": "service_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Новые значения градусов",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_app_handler.DTO_Req_DegreesUpd"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Обновленные данные",
                        "schema": {
                            "$ref": "#/definitions/internal_app_handler.DTO_Resp_UpdateDegrees"
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "Удаляет связь между гейтом и задачей",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "M-M"
                ],
                "summary": "Удалить гейт из задачи",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID задачи",
                        "name": "task_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID гейта",
                        "name": "service_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Информация об удаленной связи",
                        "schema": {
                            "$ref": "#/definitions/internal_app_handler.DTO_Resp_TaskServiceLink"
                        }
                    },
                    "400": {
                        "description": "Invalid IDs",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "internal_app_handler.DTO_Req_TaskUpd": {
            "description": "Данные для обновления описания квантовой задачи",
            "type": "object",
            "required": [
                "task_description"
            ],
            "properties": {
                "task_description": {
                    "type": "string",
                    "example": "Calculate quantum state probabilities"
                }
            }
        },
...
        "internal_app_handler.DTO_Resp_SimpleID": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
...
        "internal_app_handler.DTO_Resp_Tasks": {
            "description": "Полная информация о квантовой задаче включая гейты",
            "type": "object",
            "properties": {
                "conclusion_date": {
                    "type": "string",
                    "example": "2023-10-01T16:04:05Z"
                },
                "creation_date": {
                    "type": "string",
                    "example": "2023-10-01T15:04:05Z"
                },
                "gates_degrees": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_app_handler.DTO_Resp_GatesDegrees"
                    }
                },
                "id_task": {
                    "type": "integer",
                    "example": 1
                },
                "id_user": {
                    "type": "integer",
                    "example": 1
                },
                "res_koeff_0": {
                    "type": "number",
                    "example": 0.707
                },
                "res_koeff_1": {
                    "type": "number",
                    "example": 0.707
                },
                "task_description": {
                    "type": "string",
                    "example": "Quantum state calculation"
                },
                "task_status": {
                    "type": "string",
                    "enum": [
                        "draft",
                        "formed",
                        "completed",
                        "rejected"
                    ],
                    "example": "completed"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT токен в формате: \"Bearer {token}\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}